move b2c storefront recipe and config to recipe folder
