+++ projects/storefrontlib/src/cms-structure/page/index.ts
@@ -1,6 +1,5 @@
 export * from './component/component-wrapper.directive';
 export * from './component/page-component.module';
-export * from './config/default-cms-content.config';
 export * from './model/index';
 export * from './page-layout/page-layout.component';
 export * from './page-layout/page-layout.module';
