+++ projects/storefrontlib/src/public_api.ts
@@ -7,6 +7,7 @@ export * from './cms-pages/index';
 export * from './cms-structure/index';
 export * from './config/index';
 export * from './layout/index';
+export * from './recipes/index';
 export * from './shared/index';
 export * from './storefront-foundation.module';
 export * from './storefront.module';
