+++ projects/storefrontlib/src/recipes/config/b2c-layout-config.ts
