+++ projects/storefrontlib/src/recipes/config/index.ts
@@ -1,0 +1,2 @@
+export * from './b2c-layout-config';
+export * from './static-cms-structure/index';
