+++ projects/storefrontlib/src/recipes/config/static-cms-structure/default-cms-content.config.ts
