+++ projects/storefrontlib/src/recipes/config/static-cms-structure/default-header.config.ts
