+++ projects/storefrontlib/src/recipes/config/static-cms-structure/default-pdp-config.ts
