+++ projects/storefrontlib/src/recipes/config/static-cms-structure/index.ts
@@ -1,0 +1,3 @@
+export * from './default-cms-content.config';
+export * from './default-header.config';
+export * from './default-pdp-config';
