+++ projects/storefrontlib/src/recipes/index.ts
@@ -1,0 +1,2 @@
+export * from './b2c-storefront.module';
+export * from './config/index';
